In process of adding collaboration
To collaborate with another user, all the user specific attributes of
card where moved to card_statistics, and now it is card_statistics that
belong to the user. Cards now have many card_statistics. Copies of
topics create another card and adds in them a card_statistic that
belongs to the recipient user.